Split monolithic test file into logical per-area files
tests/test_yaml_c_wrapper.cpp (2502 lines, 106 TEST_CASEs) is replaced by
nine files grouped by concern, using issue #51's Areas list as a guide:

- test_yaml_api.cpp       low-level YAML C wrapper API
- test_expansion.cpp      parse_and_expand_PALS, forking, problem reporting
- test_correspondence.cpp build_correspondence_map
- test_key_order.cpp      map key-order preservation
- test_matching.cpp       name matching
- test_parameters.cpp     get_parameter_value / get_lattice_parameter_value
- test_expressions.cpp    evaluate_pals_expression + in-tree resolution
- test_multipass.cpp      multipass_index numbering
- test_controllers.cpp    controller expression evaluation

Shared helpers move to test_helpers.h as inline definitions (file I/O,
value/key compares, expanded-tree navigation, float compares, and the
MATCH_YAML fixture used by both matching and parameter tests). Helpers used
by a single file stay local to it. Test bodies are copied verbatim -- no
test logic changed. tests/CMakeLists.txt lists the nine sources, and the
inert CATCH_CONFIG_MAIN define is dropped (main comes from Catch2WithMain).

All 106 test cases pass (549 assertions).
